requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# LangGraph and LangChain
langgraph==0.2.28
langchain==0.3.1
langchain-core==0.3.6
langchain-openai==0.2.1
langchain-anthropic==0.2.1

# Tools and Utilities
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
typing-extensions==4.12.2

# GitHub Integration
PyGithub==2.4.0

# Code Validation
ast-tools==0.1.0

# Logging and Monitoring
python-json-logger==2.0.7

# UI Framework
streamlit==1.39.0

# Additional utilities
pygraphviz==1.13
